Move component/util.js to component/utils/StatePersistedMixin.js
Transform util into a folder capable to host multiple components in
their own file.
