+++ web/html/src/manager/audit/subscription-matching/subscription-matching-messages.js
@@ -4,7 +4,7 @@
 const React = require("react");
 const createReactClass = require('create-react-class');
 const {Table, Column, SearchField, Highlight} = require("components/table");
-const StatePersistedMixin = require("components/util").StatePersistedMixin;
+const {StatePersistedMixin} = require("components/utils/StatePersistedMixin");
 const CsvLink = require("./subscription-matching-util").CsvLink;
 const Functions = require("utils/functions");
 const Utils = Functions.Utils;
